fix : exception 생성자 추가, Test 아키텍쳐 컨셉 수정
